Added watch property
