+++ python_ta/__init__.py
@@ -285,6 +285,15 @@ def reset_linter(
     # Tuple of custom options. Note: 'type' must map to a value equal a key in the pylint/config/option.py `VALIDATORS` dict.
     new_checker_options = (
         (
+            "watch",
+            {
+                "default": False,
+                "type": "yn",
+                "metavar": "<yn>",
+                "help": "Run the HTML report server in persistent mode",
+            },
+        ),
+        (
             "pyta-number-of-messages",
             {
                 "default": 0,  # If the value is 0, all messages are displayed.
